minor fixes
Now that I have the major pieces in play for some of the cooperage items, I will start mapping out the foundation for the rest of the app. This was to get rid of a lot of unneccessary logging as well as change cooper_app.py to cooperage.py since this will be the main piece that runs everything (the cooperage, or store).