Updates And SemiFinal Script
Updated a few tweaks to step 3 and 4, and created a complete & semi-final RPS game. This is very basic, has no fluff or extra stuff. The final one may not be ready for Cisco Live 2018, but at least there's working code now. Will continue working on the final one with all the bells and whistles.